server refactor, part 2
Separate annotations from the scanpy_engine.

This introduces a new class called AnnotationsLocalFile, which stores
annotations on the local filesystem in csv format.  When we support different
ways to store annotations, those should now be easier to add in.  Also,
separating out the annotation handling from scanpy_engine will make it easier
to add in new data formats like tiledb.

There is still plenty of cleanup that can be done regarding the interactions of
the rest, adata format, and annotations.