fix static assets in hosted cxg (#1210)
The web proxy at aws eb was not finding the static assests.

The solution here is very simple: just copy the directory
containing the static assets to the top level of the artifact.zip.

This is not really the ideal solution.  According to the AWS
docs you can make a mapping to the correct location in an
an ebextentions config file.  I tried this and many combinations but
was not able to get this to work following that pattern.

Since we control the contruction of the zip file, the solution
here isn't bad, but it could probably be made better.