Introduce a config file to cellxgene
The config file format is in yaml.  The default yaml file is located
in server/common/default_config.yaml.  A user my create a yaml file
that contains a subset of these fields.  It can be used during cellxgene
launch, or for hosted cellxgene.

The code has also been refactored.  Much of the logic to check arguments
has moved from launch to app config.

It is now possible to set the tiledb context parameters using the config
file.  Other feature will soon be handled in a similar way.