Fix matrix cache race condition
The simplest fix is to acquire the cache_item lock while holding the matrix cache lock.
This potentially causes a slowdown when a new matrix is loaded for the first time.
However, even with large datasets with is hardly noticable.  If the server finds itself in
the situation where the cache is being thrashed then the solution is to increase the
cache size.

Fixes #1255