Third pass
the key of the dataroot dictionary is no longer the same as the dataroot_url.
Previously key==dataroot_url, and now those are separated.

Added an "is_multi_dataset" function to simplify logic where it branched on single vs multi.

Simplified the rest.py interface by no longer passing in the user annotations object, since
that can be retrieved from the dataset.

Fixed the eb/app.py which had not been updated with new config changes.

Added a sanity test for eb, to verify it runs and serves api requests.