Update the way the api_base_url is applied.
Previously this was done through the render_template.
The new way is using the inline scripts feature, which is a bit more flexible.

Also, add a feature to set the CORS support credentials, which seems
to be necessary for the backend/frontend separation, at least when run
locally.  This part is sort of experimental, and may be removed or changed later.