Refactor to support configurable adaptor classes
A new class can be configured to handle cxg datasets or h5ad datasets.
A type called CZI_cxg_adaptor is implemented to handle specializations needed
for interactions between explorer and portal.
For exmaple: remapping path names to dataset locations, updated corpora_props, and genesets.

So far the CZI_cxg_adaptor is just a stub.
This patch is meant to do the refactoring needed to support specialized adaptor classes.