Fix locust test.
The locust test deployment hook has been failing.
The problem was that locust is meant to just test our API and therefore requires a
limited set of modules.

However it imports a few things from server, and the modules it traverses and
the files it import, then import other pieces of the code base, that in turn bring in more than it should.

There were two main issues:
1.  the decode_matrix_FBS helper function in the server/test/unit directory has imported by locustfile.py
which pulled in many things from server/test/__init__.py.  There were a few ways to address this issue.
One was to move everything out of __init__.py to somewhere else.  This would have required a large number
of changes in the test code.  The second option was to avoid importing anything from server/test directory, which
is what I chose.  I moved the decode_matrix_FBS functionality into data_common/fbs.  This is an okay
choice, in my opinion, because there is already existing code there that decodes an FBS into a
pandas dataframe, instead of a dict.

2.  The server/__init__.py checks for plugins.  The code for this was written in the server/common/utils/utils.py
module with several other functions.  I split out the plugins handling part to its own module to reduce
dependencies.

 #2019