Copy server to local_server
This is the first step in splitting the two backends apart. local_server
will be for the desktop deployment of cellxgene while server will not be
only for hosted.

If you've been following the history of a file in local_server and would
like to keep going, you need to look at the corresponding file over in
server/